requirements.txt
# Core dependencies
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.14.0

# AI/ML dependencies
openai>=1.0.0
anthropic>=0.18.0
google-generativeai>=0.3.0
scikit-learn>=1.3.0
sentence-transformers>=2.2.0

# Google APIs
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0

# HTTP and async
aiohttp>=3.8.0
asyncio>=3.4.3
requests>=2.31.0

# NLP and text processing
nltk>=3.8.0
spacy>=3.6.0
textstat>=0.7.3

# Content extraction
beautifulsoup4>=4.12.0
lxml>=4.9.0
trafilatura>=1.6.0
readability-lxml>=0.8.1
html5lib>=1.1

# Utilities
python-dotenv>=1.0.0
validators>=0.22.0

# Export capabilities
openpyxl>=3.1.0
xlsxwriter>=3.1.0

# Caching and optimization
joblib>=1.3.0
diskcache>=5.6.0